requirements.txt
asgiref==3.7.2
astroid==3.0.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
dill==0.3.7
Django==4.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
mysqlclient==2.2.1
packaging==23.2
platformdirs==4.1.0
pycodestyle==2.11.1
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
uritemplate==4.1.1
